Make BlockedNumberBackupAgent capable of restoring from newer OS version.
Adding restoreAnyVersion removes the restriction that the OS version must
be less than or equal to that of the data being restored.  Added a check
in the onRestore method to check the backup file format version number to
ensure that it is not greater than the current supported version; if it
is we will skip restoring the data.

This means that provided the blocked number format has NOT changed,
you can restore an SDK X+1 backup file to an SDK X device.  This provides
a little bit more flexibility for users to restore from newer devices to
older devices and still retain their list of blocked numbers.

Flag: Backup and restore is not flagable; this is also destined for next
release.
Test: Added unit test to verify that we cannot restore newer backup file
versions.
Test: Manual testing was performed to ensure that restoring from newer
OS to older OS works providing the backup file format didn't change.
Bug: 317150223317053396

Change-Id: I7f09ef1809cc74f9c655d5906284b071856c6139