Make BlockedNumberProvider emergency number check more robust.
First, check to ensure TelephonyManager is non-null; it could be null if
Telephony is not present on the device at all.

Second, add "IllegalStateException" as it could be thrown if the Telephony
service is down or has crashed; this is possible for devices such as wifi
only watches which may be bridging a call via HFP (though in practice those
normally skip call filtering using `EXTRA_SKIP_CALL_FILTERING` since
call filtering isn't needed for bridged HFP calls.

Fixes: 368332015
Test: This will need to be manually tested on the specific watch model
impacted.
Test: Run unit / cts test suite for Telecom.
Flag: NONE this is a bugfix related to a provider crash.

Change-Id: I08b33fe777338651060ddab94654172682cf2582